Update fr.json
submodule: translation

So I've had to use this plugin for work (and it really helped, thanks for your hard work!), and I've noticed the french translations weren't up-to-date and didn't provide as smooth an experience as it does in english (it doesn't show the changeFrom and changeTo strings, for example). These suggested modifications aims to correct that, and also translate the few strings that weren't already translated.  Thanks for your attention! :)

Also, I'm not finding the develop branch to base the PR against, I'm only seeing the 2 included in the screenshot.